gtk: Move provider related functions outside of StyleContext
To avoid them being marked as deprecated in v4_10and is also more
sementically correct
Fixes #1317
