+++ examples/custom_editable/main.rs
@@ -17,7 +17,7 @@ fn main() -> glib::ExitCode {
     application.connect_startup(|_| {
         let provider = gtk::CssProvider::new();
         provider.load_from_data(include_str!("style.css"));
-        gtk::StyleContext::add_provider_for_display(
+        gtk::style_context_add_provider_for_display(
             &gdk::Display::default().unwrap(),
             &provider,
             800,
